Extract functions of SudokuGame to a custom hook
