+++ package.json
@@ -6,7 +6,7 @@
     "name": "Daan Rijnkels"
   },
   "license": "Apache-2.0",
-  "version": "0.8.2",
+  "version": "0.8.3",
   "private": false,
   "scripts": {
     "dev": "next dev",
